seperated bookmark class into own file
